Expose SQLiteGameDatabase publicly.
Instead of having the SQLiteGameDatabase class private in logic.cpp,
expose it in the header.  We need that for the implementation of
pending moves, so that they can access the database as well based
on the PXLogic instance.
